move utility to tools file
